requirements.txt
boto3==1.17.45
cryptography==3.4.7
pyyaml==5.4.1
